+++ api/routes/register_routes/__init__.py
@@ -8,6 +8,7 @@ from .post_organization import router as post_organization_router
 from .post_s3 import router as post_s3_router
 from .post_service import router as post_service_router
 from .post_url import router as post_url_router
+from .publish_dataset import router as publish_dataset_router
 
 router = APIRouter()
 
@@ -17,3 +18,4 @@ router.include_router(post_url_router)
 router.include_router(post_s3_router)
 router.include_router(post_service_router)
 router.include_router(post_general_dataset_router)
+router.include_router(publish_dataset_router)